Add unit test for data paging while parsing binary FTrace header
The trace data of binary FTrace files are divided into pages of the same
size. While parsing the trace header, the header parser obtains paging
information and attempts to parse the header of each data page.
Currently, there is no test to make sure that the parser reads this
information correctly. This commit adds a unit test and test data for
this feature of the parser, so that future changes will not alter the
expected behavior.

Change-Id: I3bd7e296b21c7acec35603865efd9746b9f72bab
Signed-off-by: Hoang Thuan Pham <hoang.pham@calian.ca>
Reviewed-on: https://git.eclipse.org/r/c/tracecompass.incubator/org.eclipse.tracecompass.incubator/+/191374
Tested-by: Trace Compass Bot <tracecompass-bot@eclipse.org>
Tested-by: Marco Miller <marco.miller@ericsson.com>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-by: Marco Miller <marco.miller@ericsson.com>
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/c/tracecompass.incubator/org.eclipse.tracecompass.incubator/+/191546