OTF2: Add a flows analysis with an initial view
Added a new analysis in the OTF2 plugin. One view is provided with
this analysis. This view shows the input and output flows of the
threads/processes/nodes/clusters when they are exchanging data through
MPI blocking operations. The colors of states that are displayed when
data is exchanged depends on the mean and standard deviation of all the
data that has been exchanged on the row.

Test case can be found in tracecompass-test-traces, under the name
bandwidth_issue:
tracecompass-test-traces/ctf/src/main/resources/bandwidth_issue

Signed-off-by: yoann-heitz <yoann.heitz@polymtl.ca>
Change-Id: Ie3ea1d99bb0e1d9e6f5d1646895cd2b477387702
Reviewed-on: https://git.eclipse.org/r/c/tracecompass.incubator/org.eclipse.tracecompass.incubator/+/191486
Tested-by: Marco Miller <marco.miller@ericsson.com>
Tested-by: Trace Compass Bot <tracecompass-bot@eclipse.org>
Reviewed-by: Marco Miller <marco.miller@ericsson.com>