server: Bug 579456: trace-server: use time range in TSP queries
In queries that require sampling of timestamps, replace the large
requested_times array with a requested_timerange object.

The timestamp array is then computed in the trace server.

For the tree endpoint, replace requested_times array of [start, end]
with a requested_timerange object with no nbTimes.

For backward compatibility, the computed array is then put in the query
parameters as the requested_times, overwriting it if present.

Change-Id: I4549118df6d0b298bdc8f72dbbee8a7aa3a5f8d7
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/c/tracecompass.incubator/org.eclipse.tracecompass.incubator/+/191752
Tested-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Tested-by: Trace Compass Bot <tracecompass-bot@eclipse.org>
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/c/tracecompass.incubator/org.eclipse.tracecompass.incubator/+/192558