server: Consolidate query parameters validation and conversion
Add QueryParameterUtil class with method to validate each endpoint's
query parameters map.

The method will verify the presence of required parameters and translate
parameters when necessary to conform with the data provider interfaces.

When a specific parameter is missing or invalid, identify it in the
error message response.

Remove the QueryParametersDeserializer which is no longer needed.

Make the parameters field non-null in QueryParameters and assign empty
map in constructor when not received.

Bug: 579456

Change-Id: I1cf727d5d1087e7f511fa2a7a62490b5a41d4807
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/c/tracecompass.incubator/org.eclipse.tracecompass.incubator/+/192291
Tested-by: Trace Compass Bot <tracecompass-bot@eclipse.org>
Tested-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/c/tracecompass.incubator/org.eclipse.tracecompass.incubator/+/192559