ros2: add message flow analysis
This adds a message flow analysis that can be triggered by selecting a
time graph segment in the messages view and clicking on the "follow"
action. The message flow analysis then builds a graph of the flow of the
message across the traced ROS 2 system, both forwards and backwards.

The result is shown in the message flow view, which is similar to the
messages view, but only showing what is part of the message's flow.

Screenshot: https://i.imgur.com/gZsHsxi.png
The view above is the messages view, and the view below is the message
flow view, showing the flow of the selected message, which was at the
start of the system's computation graph.

Signed-off-by: Christophe Bedard <bedard.christophe@gmail.com>
Signed-off-by: Christophe Bedard <christophe.bedard@apex.ai>
Change-Id: I0deb1b8a1ff5c7e4c925c968d373d2488c95c1d5
Reviewed-on: https://git.eclipse.org/r/c/tracecompass.incubator/org.eclipse.tracecompass.incubator/+/192551
Tested-by: Trace Compass Bot <tracecompass-bot@eclipse.org>