OTF2: Add an XY view to the flows analysis
Added an XY view similar to the one already present in the flows
analysis. This view allows to have more precise information about the
input and output flows of the different threads, processes, nodes and
clusters, while the previous timeline view allows to understand more
globally the behavior of the traced application.

When selecting several leafs in the tree viewer, the sum of the flows
values of these leafs (and then of the intermediary nodes) is then
displayed for the parents. It allows for example to compare the flows of
several processes among one node.

Signed-off-by: yoann-heitz <yoann.heitz@polymtl.ca>
Change-Id: Ifa148ab6b124e7e0a736b8038d3a109abfeddc8c
Reviewed-on: https://git.eclipse.org/r/c/tracecompass.incubator/org.eclipse.tracecompass.incubator/+/192556
Tested-by: Trace Compass Bot <tracecompass-bot@eclipse.org>
Tested-by: Marco Miller <marco.miller@ericsson.com>
Reviewed-by: Marco Miller <marco.miller@ericsson.com>
(cherry picked from commit 333ec1768a39e8cf4abf8a9171d8075036fe6d16)
Reviewed-on: https://git.eclipse.org/r/c/tracecompass.incubator/org.eclipse.tracecompass.incubator/+/195663
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>