flamechart: solve data provider resolve issue
When opening a new view from a flamechart, the uuid is appended to the
secondary id, this makes the data provider not resolved correctly.

This commit will ignore the UUID of the trace to find the provider id.

Signed-off-by: Arnaud Fiorini <fiorini.arnaud@gmail.com>
Change-Id: Ib7199bff3559199c626bdd8b9e5dc580623790c7
Reviewed-on: https://git.eclipse.org/r/c/tracecompass.incubator/org.eclipse.tracecompass.incubator/+/192597
Tested-by: Trace Compass Bot <tracecompass-bot@eclipse.org>
Tested-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
(cherry picked from commit ea4a5841f27773221d42041600d0e333aa7ce28d)
Reviewed-on: https://git.eclipse.org/r/c/tracecompass.incubator/org.eclipse.tracecompass.incubator/+/197048