server: Fix backwards compatibility of fetchTree for requested_times
The following patch updated the validation of query parameter after
introduction of requested_timerange instead of requested_times.
To be backwards compatible for clients that still use requested_times
the validation still allows requested_times to be used for now.

However, the new validation code has different behaviour that the
original implementation. This patch will align the validation to
the previous implementation.

https://git.eclipse.org/r/c/tracecompass.incubator/org.eclipse.tracecompass.incubator/+/192291

Change-Id: Ida5ca5c757512730932364240ebad57eec99b7a6
Signed-off-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/c/tracecompass.incubator/org.eclipse.tracecompass.incubator/+/193199
Tested-by: Trace Compass Bot <tracecompass-bot@eclipse.org>
Tested-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-by: Patrick Tasse <patrick.tasse@gmail.com>