OTF2: Add counter analysis
Added a counter analysis, using the generic counter aspects. To create
the different aspects, when the trace is opened, all the definitions at
the beginning of the trace are read. It allows to reconstruct the system
architecture (to build the different groups in the counter analysis) and
create the different counter aspects.

Dependencies from org.eclipse.tracecompass:
https://git.eclipse.org/r/c/tracecompass/org.eclipse.tracecompass/+/193595

Dependencies from tracecompass-test-traces:
https://git.eclipse.org/r/c/tracecompass/tracecompass-test-traces/+/193594
https://git.eclipse.org/r/c/tracecompass/tracecompass-test-traces/+/193332

Signed-off-by: yoann-heitz <yoann.heitz@polymtl.ca>
Change-Id: I8140df961c717b42e029d880306163dad9b42b6e
Reviewed-on: https://git.eclipse.org/r/c/tracecompass.incubator/org.eclipse.tracecompass.incubator/+/193333
Reviewed-by: Marco Miller <marco.miller@ericsson.com>
Tested-by: Trace Compass Bot <tracecompass-bot@eclipse.org>
Tested-by: Marco Miller <marco.miller@ericsson.com>