server: Bug 581174: Fix path handling for traces when running on Windows
When the trace server is started on Windows where the CWD
is a lower case drive letter the device handling code in
Eclipse does not work. The root cause of this is we
were making a Posix o.e.c.runtime.Path on Windows so the
c: was being identified as a path segment rather than
a device.

[Fixed] Fix path handling for traces when running server on Windows

Change-Id: Ife4f1ddd9786e298f6afd088f366d4dc148caa2d
Signed-off-by: Jonah Graham <jonah@kichwacoders.com>
Reviewed-on: https://git.eclipse.org/r/c/tracecompass.incubator/org.eclipse.tracecompass.incubator/+/197483
Tested-by: Trace Compass Bot <tracecompass-bot@eclipse.org>
Tested-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Reviewed-by: Patrick Tasse <patrick.tasse@gmail.com>