Bug 581189: Fix Flame Chart symbol resolution for LTTng UST traces
For that update the RocmTraceSymbolProviderFactory to not return a
symbol provider if the relevant ROCM analysis is not applicable for a
given trace. Also, it makes sure that the RocmTraceSymbolProvider
doesn't return a symbol if the analysis is not applicable for a given
trace (in the failing case the LTTng UST Trace).

[Fixed] Flame Chart symbol resolution for LTTng UST traces

Change-Id: I35adb2732df28e69947017856fd94022c05b95c6
Signed-off-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/c/tracecompass.incubator/org.eclipse.tracecompass.incubator/+/197573
Tested-by: Trace Compass Bot <tracecompass-bot@eclipse.org>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-by: Arnaud Fiorini <fiorini.arnaud@gmail.com>
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>