releng: Upgrade to current latest Orbit (2023-06)
Upgrade to the current latest Orbit drop R20230531010532.

Finishing upgrading from R20220302172233 to R20230302014618 solely
doesn't suffice for the goal of this change.

Adapt to the currently used versions of org.eclipse.jetty and
javax.servlet, without pinning Jetty's version anymore.

Since Orbit R20220531185310, javax.servlet is no longer provided as a
bundle. Use a Maven location to download it for now. Do the same for the
other related org.javassist, to finish building the change successfully.

Adapt both e4.20 and master target files for this. The former can be
used by means of copying [1] and [3] as [2] and [4] respectively,
locally. This change adapts [3] similarly to [4], except for the
javax.xml.bind removal which only [4] requires to build. Below [5] and
[6] show which Maven options to use for building.

Add jakarta.activation-api to the incubator.rcp feature and pom. Without
this, using the Jaeger dialog generates exceptions about related missing
classes. This happens when using the rcp binary only, so not from within
the IDE. That dialog is opened from the right-click menu on the Traces
folder while in the application.

[1]rcp/org.eclipse.tracecompass.incubator.rcp.product/legacy/tracing.incubator.product
[2]rcp/org.eclipse.tracecompass.incubator.rcp.product/tracing.incubator.product
[3]trace-server/org.eclipse.tracecompass.incubator.trace.server.product/legacy/traceserver.product
[4]trace-server/org.eclipse.tracecompass.incubator.trace.server.product/traceserver.product
[5]https://ci.eclipse.org/tracecompass/job/tracecompass-master-nightly-2021-06/configure
[6]https://ci.eclipse.org/tracecompass/job/tracecompass-incubator-nightly-2021-06/configure

Change-Id: Icc8d254c1a4b8705a36e4daa7cf7b351913a1fe2
Signed-off-by: Marco Miller <marco.miller@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/c/tracecompass.incubator/org.eclipse.tracecompass.incubator/+/202433
Tested-by: Trace Compass Bot <tracecompass-bot@eclipse.org>
Tested-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>