rocm: Rename analysis to keep compatibility before deleting
These analyses worked with traces generated by a plugin that is
superseded by another plugin that generates CTF traces that are
different in their format. Because it uses different event types
and the events are recorded differently, these analyses need to
be deprecated and deleted in the future as the previous plugin is
not maintained.

Signed-off-by: Arnaud Fiorini <fiorini.arnaud@gmail.com>
Change-Id: I3e0dc01d5e2a8dd0ab527bc055c567bd9c78f4e0
Reviewed-on: https://git.eclipse.org/r/c/tracecompass.incubator/org.eclipse.tracecompass.incubator/+/202521
Tested-by: Trace Compass Bot <tracecompass-bot@eclipse.org>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>