releng: remove javax.xml.stream from traceserver.product
Patch [1] removed javax.xml.stream from the target definition. The
incubator build was still successful, because that library was provided
through the Trace Compass update site. However, later on when it was
removed from the trace compass update site through patch [2] then the
incubator build started failing.

This patch removes it from the traceserver.product file as well. It is
not needed anymore and trace server works as expected.

[1] https://git.eclipse.org/r/c/tracecompass.incubator/org.eclipse.tracecompass.incubator/+/202433
[2] https://git.eclipse.org/r/c/tracecompass/org.eclipse.tracecompass/+/202642

Change-Id: Ib6b2bfca5dfdc851e58a747350c08bb9376e824d
Signed-off-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/c/tracecompass.incubator/org.eclipse.tracecompass.incubator/+/202891
Tested-by: Trace Compass Bot <tracecompass-bot@eclipse.org>
Tested-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-by: Patrick Tasse <patrick.tasse@gmail.com>