swagger: Move @OpenAPIDefinition in own class
Before this patch, the application level swagger @OpenAPIDefinition was
arbitrarily added to the DataProviderService class, because only one
class can specify that. This patch creates a new class TcOpenApiResource
that extends the class OpenApiResource that is used to enable the
swagger end-point. With this new class a central class is used for
application level swagger definitions.

Change-Id: Ie42e9614ebe80ad58339d79d664f17b692f77539
Signed-off-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/c/tracecompass.incubator/org.eclipse.tracecompass.incubator/+/204787
Tested-by: Patrick Tasse <patrick.tasse@gmail.com>
Tested-by: Trace Compass Bot <tracecompass-bot@eclipse.org>
Reviewed-by: Patrick Tasse <patrick.tasse@gmail.com>