execution comparison: second commit
Adding the option of textual input of dates and query
to build the differential flame graph.

An "EventDensityView" is used to select a time range for each group.
Select the desired traces in an experiment.

Show the difference ratio based on self time required a modification to
AggregatedCalledFunction.

Introduce "ParametricWeightedTreeUtils" to get "SelfTime" instead of "duration"
while building the differential Trees.

TODO:

* add grouping strategies (WALL TIME) Not Valid: It was
  Discussed for Async Trace Comparison
* move grouping to hamburger menu Not yet
* add progress bar Not yet
* instrument code (pretty good)
* add tests Not yet
* Make work with mainline (should be a simple ID swap) Not yet
* Make a registry for tracetype->callstack ID Not yet
* remove hard disk operations from UI thread (in progress) Not yet
* Isolate duplicating flame graph: Done
* Add way to select the entire trace easily: Done
* Add way to manually select precise time range (import from main?) :Done

Signed-off-by: Fariba Daneshgar <faraji.daneshgar@gmail.com>