execution comparison: adding reset button and some modifications
Introduce a module to compare two groups of executions using a
differential flame graph.

An "EventDensityView" is used to select a time range for each group.
Select the desired traces in an experiment.

Showing the difference ratio based on self time (it  required a modification to
AggregatedCalledFunction).

Introducing the "ParametricWeightedTreeUtils" to get "SelfTime" instead of "duration"
while building the differential Trees.

Isolating duplicating flame graph

TODO:

* add grouping strategies (WALL TIME) Not Valid: It was
  Discussed for Async Trace Comparison
* move grouping to hamburger menu
* add progress bar
* instrument code (pretty good)
* add tests
* Make work with mainline (should be a simple ID swap)
* Make a registry for tracetype->callstack ID
* remove hard disk operations from UI thread (in progress)
* Add way to select the entire trace easily:
* Add way to manually select precise time range (import from main?)
* query will be added

Signed-off-by: Fariba Daneshgar <faraji.daneshgar@gmail.com>
Signed-off-by: Matthew Khouzam <matthew.khouzam@ericsson.com>