execution-comparison: improve code
This commit fixes various types of formatting issues and code warnings
coming from `Eclipse` and `SonarLint`. The code is also refactored in
various places to reduce the complexity of certain methods.

In addition, it also fixes a bug
where the time range panel (from and to) is broken when initially
launching TC.

Finally, it adds clearer labels for the Y-Axis (Event Count instead of Y
Axis).

Signed-off-by: Vlad Arama <vlad.arama@ericsson.com>