dpdk: add packet throughput analysis for ethdev
Introduces packet throughput analysis for DPDK applications
using the ethdev library. The analysis calculates packet throughput in
both bits per second (bps) and packets per second (pps) for transmitted
and received Ethernet traffic on a per-queue basis.

Note that the calculation of throughput in bps requires the existence of
custom profiling events in the trace

Signed-off-by: Adel Belkhiri <adel.belkhiri@gmail.com>