dpdk: add packet distribution statistics analysis
Introduce packet distribution statistics analysis for PMD threads in the ethdev
library. This analysis calculates various statistics related to the
distribution of packets retrieved in a single rte_eth_rx_burst() call, on a
per-thread and per-queue basis. The computed statistics include the minimum,
maximum, average number of packets retrieved, as well as the standard deviation.

Signed-off-by: Adel Belkhiri <adel.belkhiri@gmail.com>