[trace server] update link to Jetty SSL setup doc
The existing link [1], to the Jetty documentation for the SSL
setup, no longer resolves to the correct page, instead leading
to a root documentation page on jetty.org, where one needs to
pick the Jetty version to get to the documentation, and then
search for the relevant SSL-related page.

Under there, I found what seems to be the relevant link for SSL
documentation [2], for Jetty version 12 (latest as of now).

[1]: https://www.eclipse.org/jetty/documentation/current/configuring-ssl.html
[2]: https://jetty.org/docs/jetty/12/operations-guide/keystore/index.html

Signed-off-by: Marc Dumais <marc.dumais@ericsson.com>