releng: Bundle java runtime in Trace Compass RCP and Trace Server RCP
Update target files to include JustJ JREs https://eclipse.dev/justj/
with respective version for that release. Use latest available JRE
version for respective major version.

This will make it easier for end-users to use Trace Compass because Java
doesn't need to be installed separately and it will include the correct,
required version.

Notes:

- The stripped version of the JRE is used with debug information
stripped and excluding the src.jar with sources used only for debugging.
This minimizes the size of the JRE package because it's less than 1/2
the size of the corresponding full version. If you'd like to debug the
RCP (e.g. using remote debugging), use a separated JVM as described
below.
- To run the RCP with a different JVM (e.g. installed on user's laptop),
open the tracecompass.ini/tracecompass-server.ini file and modify the
path settings for -vm. Or remove it completely then the default JRE of
the computer is used.

Use a newer update site for help-docs-eclipserun-repo in main pom.xml to
allow builds be successful when building with Java 21.

Signed-off-by: Bernd Hufmann <bernd.hufmann@ericsson.com>