swagger: Allow double and remove boolean in OutputElementStyle
Previously only float values were allowed. Don't restrict to float and
allow Double as well. Note when deserializing JSON strings a floating
point value will be deserialized to Double, e.g. when using Jackson
deserializer (Java) or in Javascript.

Remove boolean because it can be expressed as String and boolean have
no equivalent type in CSS where the StyleProperties are inspired of.

Contributes to:
https://github.com/eclipse-cdt-cloud/trace-server-protocol/issues/102

Signed-off-by: Bernd Hufmann <bernd.hufmann@ericsson.com>