server: Introduce ErrorResponse and use it everywhere
When an error is created when processing a client request, return
an ErrorResponse JSON object that includes
- title: short, human-readable description of the error
- detail: optional, human-readable explanation of the error

This standardizes the error response to common data structure. The
field names and purpose are inspired by RFC 9457.

For 409 errors (trace or experiment) return extended ErrorResponse
containing the conflicting trace or experiment respectively.

This commit also updates the swagger documentation in respect to the
ErrorResponses accordingly.

This commit also fixes some incorrect or missing the swagger
definitions.

Add description to Trace and Experiment model so that they are not
overridden by the experiment/trace error description.

Contributes to fix issue:

https://github.com/eclipse-cdt-cloud/trace-server-protocol/issues/122

Signed-off-by: Bernd Hufmann <bernd.hufmann@ericsson.com>