ros2: store instrumentation version per-process, not per-trace (#207)
ROS 2 traces include the version of the ROS package used to generate the
trace, which effectively gives us a version number for the
instrumentation. This helps us interpret the trace events and their
fields.

We are reading this version number from ROS 2 traces and are currently
storing it at the trace level. However, this version number does not
really belong to the trace: it belongs to the process itself.

This flaw became evident when trying to combine 2 traces of the same
ROS 2 system/processes:

1. Trace that includes trace events from the start of the application,
   which includes the version information
2. Trace that includes trace events from later in the application's
   lifetime, which does not include the version information
    * This is because the version information is currently collected
      through a tracepoint triggered during the initialization of ROS 2

Therefore, we need to associate the version information to the process.
In the case presented above, since the processes being traced by both
traces are the same, the same version information will be shared.

Do this by storing the version number in the Ros2Objects state system as
a `{ process -> version number }` map.

Signed-off-by: Christophe Bedard <bedard.christophe@gmail.com>