server: Fix SeriesModel xValues swagger definition/serializers/tests
Use xValues for array of longs for x values, e.g. timestamps
Use xCategoires for array of category strings for x values
use XRanges for arrays of range (start/stop) for x values

This simplifies the swagger definition and implemntation for xValues of
SeriesModel for both server and client side.

Signed-off-by: Bernd Hufmann <bernd.hufmann@ericsson.com>