server.tests: Regenerate Java TSP client and XY fix tests
Fix for #236 required a regeneration of the Java TSP client and caused
changes in the API that required the XY tests to be fixed.

Remove SamplingSerializer and SamplingSerializerTest because it's part
of SeriesModelSerializer and SeriesModelSerializerTest respectively.

Set flag failOnUnknownProperties when generating Java TSP client which
helps with detecting correct class for deserialization especially for
oneOf cases, e.g. oneOf AxisDomainCategorical, AxisDomainRange.

Set hideGenerationTimestamp flag to true since generation timestamp can
be derived from the git history.

Signed-off-by: Bernd Hufmann <bernd.hufmann@ericsson.com>